refactor SimpleItem as TreeItemLayout
