+++ apps/app/src/components/TreeItem/TreeItemLayout.module.scss
@@ -1,5 +1,5 @@
 // show / hide on hover
-.simple-item {
+.tree-item-layout {
   :global {
     .list-group-item {
       &:hover {
